# speedup IN-MH parser