--- a/parsers/SG.py
+++ b/parsers/SG.py
@@ -6,8 +6,11 @@ import re
 
 import arrow
 from PIL import Image
+from pytessy import pytessy
 from pytesseract import image_to_string
 import requests
+
+from parsers.IN_MH import RGBtoBW
 
 TIMEZONE = 'Asia/Singapore'
 
@@ -68,11 +71,9 @@ def get_solar(session, logger):
 
     w, h = solar_image.size
     text_only = solar_image.crop((w*0.53125, h*0.74375, w*0.9296875, h*0.91875))
-    gray = text_only.convert('L')
-    threshold_filter = lambda x: 0 if x < 10 else 255
-    black_white = gray.point(threshold_filter, '1')
-
-    text = image_to_string(black_white, lang='eng')
+    black_white = RGBtoBW(text_only)
+    digit_reader = pytessy.PyTessy(language='eng')
+    text = digit_reader.read(black_white.tobytes(), black_white.width, black_white.height, 1)
 
     try:
         pattern = r'Est. PV Output: (.*)MWac'