--- a/parsers/requirements.txt
+++ b/parsers/requirements.txt
@@ -18,3 +18,4 @@ testfixtures==6.0.0
 opencv-python==4.2.0.32
 xlrd==1.1.0
 freezegun==0.3.15
+pytessy==0.1.0