--- a/pyproject.toml
+++ b/pyproject.toml
@@ -29,6 +29,7 @@ opencv-python = {version="4.2.0.32", optional=true}
 xlrd = {version="1.2.0", optional=true}
 freezegun = {version="0.3.15", optional=true}
 signalr-client-threads = {version="0.0.12", optional=true}
+pytessy = {version="^0.1.0", optional=true}
 
 [tool.poetry.dev-dependencies]
 click = "6.7"
@@ -61,7 +62,8 @@ parsers = [
     "opencv-python",
     "xlrd",
     "freezegun",
-    "signalr-client-threads"
+    "signalr-client-threads",
+    "pytessy"
 ]
 
 [build-system]