Using pytessy sped up the pytesseract portion of the code by ~10% on my machine. This is linked to issue #2758
I haven't incorporated it into the other parsers that use pytesseract yet as it was causing an error in JP_KN (OSError: [WinError 127] The specified procedure could not be found). Narrowed this down to the importing of beautifulsoup4

@GitMatze

On my machine:

Tesseract:
%timeit pytesseract.image_to_string(imgs_line, lang='digits_comma', config='--psm 7')
521 ms ± 7.66 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

pytessy:
%timeit digit_reader.read(imgs_line.tobytes(), imgs_line.size[0], imgs_line.size[1], 1)
484 ms ± 20.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

and for a file taken at an earler time:
Tesseract:
%timeit pytesseract.image_to_string(imgs_line, lang='digits_comma', config='--psm 7')
1.02 s ± 3.97 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

pytessy:
%timeit digit_reader.read(imgs_line.tobytes(), imgs_line.size[0], imgs_line.size[1], 1)
939 ms ± 5.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)