@GitMatze 2020-10-29
@ssmssam  Seems like you can't pass the 'psm 7' argument with pytessy? It advises tesseract to read the image as a single text line. I will do some logging to see if there are any differences in prediction reliability.