@ssmssam 2020-11-03
Note, I have also made a change to the parser SG.py to help with issue #2779 . I can seperate this into a different pull request if that is easier @GitMatze  .