@ssmssam 2021-02-25
That's ok @skovhus  . Sorry, I didn't think to try it out on Linux before. Will try that next time round. I haven't got any off the top of my head but will have a look around and see if I can find something!